+++ 1.8.9-Forge/src/main/java/net/ccbluex/liquidbounce/api/minecraft/enchantments/IEnchantment.kt
@@ -7,7 +7,7 @@
 package net.ccbluex.liquidbounce.api.minecraft.enchantments
 
 interface IEnchantment {
-    val effectId: Short
+    val effectId: Int
 
     fun getTranslatedName(level: Int): String
 }
