+++ 1.8.9-Forge/src/main/java/net/ccbluex/liquidbounce/api/minecraft/util/IFoodStats.kt
@@ -7,5 +7,5 @@
 package net.ccbluex.liquidbounce.api.minecraft.util
 
 interface IFoodStats {
-    val foodLevel: Float
+    val foodLevel: Int
 }
